% Generated by roxygen2 (4.0.0): do not edit by hand
\name{focus}
\alias{focus}
\title{Perform focus step of two-stage GSA}
\usage{
focus(data, labels, genesets, sigSets, genes = rownames(data), B = 100,
  test = "wilcoxon", side = "abs", adjust = "WY", reuse.p = F)
}
\arguments{
\item{data}{Matrix containing expression values. Genes are in rows, samples
in columns.}

\item{labels}{Vector containing gene identifiers that correspond to those in
genesets}

\item{genesets}{List of genesets, given as vectors of gene identifiers}

\item{sigSets}{Output from screening step}

\item{genes}{Vector with gene identifier - defaults to rownames}

\item{B}{Number of resampling iterations (e.g. for Westfall & Young)}

\item{test}{name of test function, if reuse.p is FALSE a test function that
given if reuse.p is TRUE a function that given takes a formula argument
(x~labels) and returns an object with $p.value slot. Vector of gene
identifiers returns the corresponding unadjusted p-values}

\item{side}{should one or two-sided tests be performed}

\item{adjust}{multiplicity adjustment to use}

\item{reuse.p}{Sometimes it is desired to re-use the p-values of a previous
differential gene expression analysis. If TRUE test should be a vector of
unadjusted p-values}
}
\description{
Performs focus step of two-stage GSA. In this step single genes are tested
at significance levels resulting from the screening step, which only tests
gene sets.
}
\author{
Florian Klinglmueller
}

