% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{reportFocus}
\alias{reportFocus}
\title{Produces a nice report}
\usage{
reportFocus(scr, pval, fc, q, linebreak = 75, pw.annotation, gene.annotation,
  pw.kv, gene.kv)
}
\arguments{
\item{scr}{output from screening step}

\item{pval}{}

\item{fc}{}

\item{q}{}

\item{linebreak}{}

\item{pw.annotation}{Annotation environment that maps pathway identifiers to
some annotation}

\item{gene.annotation}{Annotation environment that maps gene identifiers to
some annotation}

\item{pw.kv}{Character value of the form c('KEY','VALUE') specifying which
key- and valuetypes should be used in the corresponding annotation database}

\item{gene.kv}{Character value of the form c('KEY','VALUE') specifying which
key- and valuetypes should be used in the corresponding annotation database
w}
}
\description{
Produces a nice report
}
\author{
Florian Klinglmueller
}

