% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{reportScreening}
\alias{reportScreening}
\title{Produces human readable summary report}
\usage{
reportScreening(scr, pw.annotation, pw.kv)
}
\arguments{
\item{scr}{Output object from screening step of twoStageGSA}

\item{pw.annotation}{Annotation environment that maps pathway identifiers to
some annotation}

\item{pw.kv}{Character value of the form c('KEY','VALUE') specifying which
key- and valuetypes should be used in the corresponding annotation database}
}
\description{
Produces human readable summary report
}
\author{
Florian Klinglmueller
}

