% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{twoStageGSA-package}
\alias{twoStageGSA-package}
\title{Two stage gene set analysis

Methods to apply two stage - screening and focus - gene set analysis controlling a number of false discovery rate derived error rates}
\description{
Two stage gene set analysis

Methods to apply two stage - screening and focus - gene set analysis controlling a number of false discovery rate derived error rates
}

