% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{screening}
\alias{screening}
\title{Performs the screening step}
\usage{
screening(data, labels, genesets, genes = rownames(data), B = 0, q = 0.05,
  settest = "globalTest", min = 6, max = 100, parcomp = c("none",
  "parallel", "bt88.03.704"))
}
\arguments{
\item{data}{Expression values as a matrix with genes in rows}

\item{labels}{Object defining the group labels of the dataset. E.g. for
globalTest this may be a factor.}

\item{genesets}{A list whose elements are vectors of gene identifiers
representing the gene sets.}

\item{genes}{vector of gene identifiers specifying the gene each row of data
corresponds to and that match to the identifyers used to define gene sets.
If not given rownames are assumed to be gene identifyers.}

\item{B}{Number of randomizations to use if a randomization test is used}

\item{q}{q-Value cutoff for selection}

\item{settest}{A character string giving the name of the screening test
procedure. Either one of the methods already implemented in the package, or
a custom test function (see Details).}

\item{min}{minimum number of genes per set for which measurements are available}

\item{max}{maximum number of genes per set for which measurements are available}

\item{parcomp}{should parallel processing be used and which implementation of mclapply}
}
\description{
twoStageGSA provides a couple of screening test procedures. Those
are 'globalTest', 'invNormal', 'hotellingT', and
'nettleton'. Custom functions may be used. In that case settest has
to point to a function with the name matching
paste('screening_',settest). The corresponding function needs to
return an unadjusted p-value for a given geneset. See
\code{\link{screening-tests}} and the vignette for further details
and some examples.
}

