% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageGSA-package.R
\docType{package}
\name{twoStageGSA-package}
\alias{twoStageGSA-package}
\title{Two stage gene set analysis

Methods to apply two stage - screening and focus - gene set analysis controlling a number of false discovery rate derived error rates}

